deleted space fram since spacechat was doing all the functionality anyways
